+++ hardhat.config.ts
@@ -52,7 +52,7 @@ const config: HardhatUserConfig = {
       // timeout: 150000,
       forking: {
         url: `https://eth-mainnet.alchemyapi.io/v2/${ALCHEMY_ID}`,
-        blockNumber: 16347107,
+        blockNumber: 16369630,
       },
     },
     mainnet: {
