+++ src/addresses.ts
@@ -30,7 +30,7 @@ export const getAddresses = (network: string): Addresses => {
         ArrakisDevAdmin: "",
         ArrakisDevOwner: "0xAa2E0c5c85ACb7717e58060AB3c96d2B184EE07C",
         ArrakisFeeTreasury: "0x2FF5D1da4985113F467BBBFF015e76ce8aB05F29",
-        OneInchRouter: "0x1111111254fb6c44bac0bed2854e76f90643097d",
+        OneInchRouter: "0x1111111254EEB25477B68fb85Ed929f73A960582",
         SwapRouter: "0xE592427A0AEce92De3Edee1F18E0157C05861564",
         WETH: "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
         DAI: "0x6B175474E89094C44Da98b954EedeAC495271d0F",
@@ -49,7 +49,7 @@ export const getAddresses = (network: string): Addresses => {
         ArrakisDevAdmin: "",
         ArrakisDevOwner: "0xAa2E0c5c85ACb7717e58060AB3c96d2B184EE07C",
         ArrakisFeeTreasury: "0x2FF5D1da4985113F467BBBFF015e76ce8aB05F29",
-        OneInchRouter: "0x1111111254fb6c44bac0bed2854e76f90643097d",
+        OneInchRouter: "0x1111111254EEB25477B68fb85Ed929f73A960582",
         SwapRouter: "0xE592427A0AEce92De3Edee1F18E0157C05861564",
         WETH: "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
         DAI: "0x6B175474E89094C44Da98b954EedeAC495271d0F",
