+++ src/swapAndAddTest.ts
@@ -1,23 +1,21 @@
 import { expect } from "chai";
-import { ethers, network, deployments } from "hardhat";
+import { ethers, network } from "hardhat";
 import {
   ArrakisV2RouterExecutor,
   ArrakisV2GenericRouter,
   SwapResolver,
   ERC20,
-  ManagerMock,
+  ArrakisV2,
 } from "../typechain";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
-import { Addresses, getAddresses } from "./addresses";
 import {
   swapTokenData,
   quote1Inch,
   mockPayloads,
   OneInchDataType,
 } from "./oneInchApiIntegration";
-import { BigNumber, ContractTransaction, Contract, Signer } from "ethers";
-import UniswapV3Factory from "@uniswap/v3-core/artifacts/contracts/UniswapV3Factory.sol/UniswapV3Factory.json";
-import UniswapV3Pool from "@uniswap/v3-core/artifacts/contracts/UniswapV3Pool.sol/UniswapV3Pool.json";
+import { BigNumber, ContractTransaction, Contract } from "ethers";
+import { Addresses, getAddresses } from "./addresses";
 
 const addresses: Addresses = getAddresses(network.name);
 
@@ -30,7 +28,7 @@ export const swapAndAddTest = async (
 
   resolver: Contract,
 
-  vault: Contract,
+  vault: ArrakisV2,
   token0: ERC20,
   token1: ERC20,
   rakisToken: ERC20,
@@ -55,11 +53,8 @@ export const swapAndAddTest = async (
   const decimalsToken1 = await token1.decimals();
   amount0Max = ethers.utils.parseUnits(amount0Max.toString(), decimalsToken0);
   amount1Max = ethers.utils.parseUnits(amount1Max.toString(), decimalsToken1);
-  // console.log("amount0Max: ", amount0Max.toString());
-  // console.log("amount1Max: ", amount1Max.toString());
 
   const isToken0Weth: boolean = token0.address === addresses.WETH;
-  // console.log("isToken0Weth: ", isToken0Weth);
 
   // amounts used for getMintAmounts(), to be filled later depending on swap amounts
   let amount0Use: BigNumber;
@@ -81,19 +76,11 @@ export const swapAndAddTest = async (
   const balanceStRakisBefore = stRakisToken
     ? await stRakisToken.balanceOf(signerAddress)
     : ethers.BigNumber.from(0);
-  // console.log("balance0Before: ", balance0Before?.toString());
-  // console.log("balance1Before: ", balance1Before?.toString());
-  // console.log("balanceEthBefore: ", balanceEthBefore?.toString());
 
   // we store working payloads from 1inch API for the swaps needed for tests and block number tests are pinned to
   let swapParams: OneInchDataType;
   let swapAmountIn: BigNumber;
   let swapAmountOut: BigNumber;
-
-  // const vaultNamet = (await token0.symbol()) + "/" + (await token1.symbol());
-
-  // console.log("vaultName : ", vaultNamet);
-  // console.log("scenario : ", mockPayloadScenario);
 
   if (mockPayloadScenario && shouldUseMockPayloads) {
     const vaultName = (await token0.symbol()) + "/" + (await token1.symbol());
@@ -122,7 +109,7 @@ export const swapAndAddTest = async (
 
     const chainID =
       network.name == "hardhat"
-        ? "137"
+        ? "1"
         : network.config.chainId?.toString() ?? "1";
 
     // amount here is not so important, as what we want is an initial price for this asset pair
@@ -146,7 +133,6 @@ export const swapAndAddTest = async (
     const priceX18 = numerator
       .mul(ethers.utils.parseEther("1"))
       .div(denominator);
-    // console.log("price check:", priceX18.toString());
 
     // given this price and the amounts the user is willing to spend
     // which token should be swapped and how much
@@ -157,11 +143,6 @@ export const swapAndAddTest = async (
       amount1Max,
       priceX18
     );
-    // console.log(
-    //   "calculateSwapAmount - result.swapAmount.toString(): ",
-    //   result.swapAmount.toString()
-    // );
-    // console.log("calculateSwapAmount - result.zeroForOne: ", result.zeroForOne);
     expect(result.zeroForOne).to.be.equals(zeroForOne);
 
     // now that we know how much to swap, let's get a new quote
@@ -171,7 +152,6 @@ export const swapAndAddTest = async (
       zeroForOne ? token1.address : token0.address,
       result.swapAmount.toString()
     );
-    console.log("quoteAmount2:", quoteAmount2);
 
     const numerator2 = ethers.BigNumber.from(quoteAmount2).mul(
       zeroForOne
@@ -186,7 +166,6 @@ export const swapAndAddTest = async (
     const price2 = numerator2
       .mul(ethers.utils.parseEther("1"))
       .div(denominator2);
-    // console.log("price2 check:", price2.toString());
 
     // given the new price, let's get a new swap amount
     const result2 = await swapResolver.calculateSwapAmount(
@@ -195,10 +174,6 @@ export const swapAndAddTest = async (
       amount1Max,
       price2
     );
-    // console.log(
-    //   "calculateSwapAmount - result2.swapAmount.toString():",
-    //   result2.swapAmount.toString()
-    // );
     expect(result2.zeroForOne).to.be.equals(zeroForOne);
 
     // given this new swapAmount, how much of the other token will I receive?
@@ -208,7 +183,6 @@ export const swapAndAddTest = async (
       zeroForOne ? token1.address : token0.address,
       result2.swapAmount.toString()
     );
-    // console.log("quoteAmount3:", quoteAmount3);
 
     swapAmountIn = result2.swapAmount;
     swapAmountOut = ethers.BigNumber.from(quoteAmount3);
@@ -221,41 +195,17 @@ export const swapAndAddTest = async (
       routerExecutor.address,
       slippage.toString()
     );
-  }
 
-  // now that we have swapData, calculate amounts used for getMintAmounts()
-  // if (zeroForOne) {
-  //   amount0Use = amount0Max.sub(swapAmountIn);
-  //   amount1Use = amount1Max.add(swapAmountOut);
-  // } else {
-  //   amount0Use = amount0Max.add(swapAmountOut);
-  //   amount1Use = amount1Max.sub(swapAmountIn);
-  // }
-  // console.log("amount0Use.toString(): ", amount0Use.toString());
-  // console.log("amount1Use.toString() ", amount1Use.toString());
-  // const mintAmounts = await resolver.getMintAmounts(
-  //   vault.address,
-  //   amount0Use,
-  //   amount1Use
-  // );
-  // console.log(
-  //   "mintAmounts.amount0.toString() ",
-  //   mintAmounts.amount0.toString()
-  // );
-  // console.log(
-  //   "mintAmounts.amount1.toString() ",
-  //   mintAmounts.amount1.toString()
-  // );
-  // console.log(
-  //   "mintAmounts.mintAmount.toString() ",
-  //   mintAmounts.mintAmount.toString()
-  // );
+    // log these to store payload after
+    console.log("swapAmountIn: ", swapAmountIn.toString());
+    console.log("swapAmountOut: ", swapAmountOut.toString());
+    console.log("swapParams: ", swapParams);
+  }
 
   // calculate minimum amount out on the swap considering slippage passed
   const amountOut = swapAmountOut
     .mul(ethers.BigNumber.from((100 - slippage).toString()))
     .div(ethers.BigNumber.from((100).toString()));
-  // console.log("amountOut.toString(): ", amountOut.toString());
 
   if (!rebalance) {
     rebalance = false;
@@ -282,22 +232,15 @@ export const swapAndAddTest = async (
     swapRouter: swapParams.to,
     swapPayload: swapParams.data,
 
-    userToRefund: "0x0000000000000000000000000000000000000000",
+    userToRefund: signerAddress,
   };
   const swapAndAddData = {
     addData: addData,
     swapData: swapData,
   };
 
-  // console.log("swapIn : ", swapAmountIn.toString());
-  // console.log("swapOut : ", amountOut.toString());
-  // console.log("to : ", swapParams.to);
-  // console.log("data : ", swapParams.data);
-
-  // flag indicating if "Swapped" event fired
-  let hasSwapped = false;
-  // flag indicating if "Minted" event fired
-  let hasMinted = false;
+  let hasSwapped = false; // flag indicating if "Swapped" event fired
+  let hasMinted = false; // flag indicating if "Minted" event fired
 
   // object to be filled with "Swapped" event data
   const swapppedEventData = {
@@ -328,7 +271,7 @@ export const swapAndAddTest = async (
 
   // listener for getting data from "Minted" event
   vault.on(
-    "Minted",
+    "LogMint",
     (
       receiver: string,
       mintAmount: BigNumber,
@@ -357,6 +300,7 @@ export const swapAndAddTest = async (
 
   let swapAndAddTxPending: ContractTransaction;
 
+  // call swapAndAddLiquidity
   if (useETH) {
     if (isToken0Weth) {
       const value = transactionEthValue || swapAndAddData.addData.amount0Max;
@@ -402,8 +346,10 @@ export const swapAndAddTest = async (
     }
   }
 
+  // wait for tx
   const swapAndAddTx = await swapAndAddTxPending.wait();
 
+  // calculate eth spent in tx
   const ethSpentForGas = swapAndAddTx.gasUsed.mul(
     swapAndAddTx.effectiveGasPrice
   );
@@ -444,16 +390,6 @@ export const swapAndAddTest = async (
   // calculate expected refunds
   const refund0 = amount0Use.sub(mintedEventData.amount0In);
   const refund1 = amount1Use.sub(mintedEventData.amount1In);
-
-  // console.log("amount0Use: ", amount0Use.toString());
-  // console.log("amount1Use: ", amount1Use.toString());
-  // console.log("refund0: ", refund0.toString());
-  // console.log("refund1: ", refund1.toString());
-  // console.log("balanceEthAfter: ", balanceEthAfter.toString());
-  // console.log("balanceEthBefore: ", balanceEthBefore.toString());
-  // console.log("swapData.amount0Max: ", swapData.amount0Max.toString());
-  // console.log("swapData.amount1Max: ", swapData.amount1Max.toString());
-  // console.log("ethSpentForGas: ", ethSpentForGas.toString());
 
   // validate balances
   if (!useETH) {
@@ -541,203 +477,8 @@ export const swapAndAddTest = async (
     expect(genericRouterBalanceStRakis).to.equal(ethers.constants.Zero);
   }
 
+  // validate we cannot mint with amounts refunded
   await expect(
     resolver.getMintAmounts(vault.address, refund0, refund1)
   ).to.be.revertedWith("ArrakisVaultV2: mint 0");
-};
-
-export const getPeripheryContracts = async (
-  owner: Signer
-): Promise<[SwapResolver, ArrakisV2RouterExecutor, ArrakisV2GenericRouter]> => {
-  // getting resolver contract
-  const resolverAddress = (await deployments.get("SwapResolver")).address;
-  const swapResolver = (await ethers.getContractAt(
-    "SwapResolver",
-    resolverAddress
-  )) as SwapResolver;
-
-  // getting router executor contract
-  const routerExecutorAddress = (
-    await deployments.get("ArrakisV2RouterExecutor")
-  ).address;
-  const routerExecutor = (await ethers.getContractAt(
-    "ArrakisV2RouterExecutor",
-    routerExecutorAddress
-  )) as ArrakisV2RouterExecutor;
-
-  // getting generic router contract
-  const genericRouterAddress = (await deployments.get("ArrakisV2GenericRouter"))
-    .address;
-  const genericRouter = (await ethers.getContractAt(
-    "ArrakisV2GenericRouter",
-    genericRouterAddress
-  )) as ArrakisV2GenericRouter;
-
-  // updating genericRouter's executor
-  await genericRouter
-    .connect(owner)
-    .updateRouterExecutor(routerExecutor.address);
-
-  return [swapResolver, routerExecutor, genericRouter];
-};
-
-export const getManagerMock = async (): Promise<ManagerMock> => {
-  const managerAddress = (await deployments.get("ManagerMock")).address;
-  const managerMock = (await ethers.getContractAt(
-    "ManagerMock",
-    managerAddress
-  )) as ManagerMock;
-  return managerMock;
-};
-
-export const deployArrakisV2 = async (
-  signer: SignerWithAddress,
-  token0Address: string,
-  token1Address: string,
-  fee: number,
-  resolver: Contract,
-  managerAddress: string
-): Promise<[Contract]> => {
-  const signerAddress = await signer.getAddress();
-
-  // getting vault factory
-  const vaultV2Factory = await ethers.getContractAt(
-    "IArrakisV2Factory",
-    addresses.ArrakisV2Factory,
-    signer
-  );
-
-  // getting uniswap factory
-  const uniswapV3Factory = new ethers.Contract(
-    addresses.UniswapV3Factory,
-    UniswapV3Factory.abi,
-    signer
-  );
-
-  // getting pool address
-  const poolAddress = await uniswapV3Factory.getPool(
-    token0Address,
-    token1Address,
-    fee
-  );
-
-  // getting uniswap pool
-  const uniswapV3Pool = new ethers.Contract(
-    poolAddress,
-    UniswapV3Pool.abi,
-    signer
-  );
-
-  // getting pool data
-  const slot0 = await uniswapV3Pool.slot0();
-  const tickSpacing = await uniswapV3Pool.tickSpacing();
-  const lowerTick = slot0.tick - (slot0.tick % tickSpacing) - tickSpacing;
-  const upperTick = slot0.tick - (slot0.tick % tickSpacing) + 2 * tickSpacing;
-
-  // get initial amounts
-  const res = await resolver.getAmountsForLiquidity(
-    slot0.tick,
-    lowerTick,
-    upperTick,
-    ethers.utils.parseUnits("1", 18)
-  );
-
-  // deploying vault
-  const tx = await vaultV2Factory.deployVault(
-    {
-      feeTiers: [fee],
-      token0: token0Address,
-      token1: token1Address,
-      owner: signerAddress,
-      init0: res.amount0,
-      init1: res.amount1,
-      manager: managerAddress,
-      routers: [addresses.SwapRouter],
-      burnBuffer: 1000,
-    },
-    true
-  );
-
-  const rc = await tx.wait();
-  const event = rc?.events?.find(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (event: any) => event.event === "VaultCreated"
-  );
-  const result = event?.args;
-
-  // getting vault
-  const vault = await ethers.getContractAt("IArrakisV2", result?.vault, signer);
-
-  return [vault];
-};
-
-export const getFundsFromFaucet = async (
-  faucetAddress: string,
-  token: ERC20,
-  targetAddress: string
-) => {
-  if (network.name === "local") {
-    const provider = new ethers.providers.JsonRpcProvider(
-      "http://localhost:8545"
-    );
-    await provider.send("hardhat_impersonateAccount", [faucetAddress]);
-    await provider.send("hardhat_setBalance", [
-      faucetAddress,
-      "0x313030303030303030303030303030303030303030",
-    ]);
-    const faucetSigner = await provider.getSigner(faucetAddress);
-    const faucetBalance = await token.balanceOf(faucetAddress);
-    await token.connect(faucetSigner).transfer(targetAddress, faucetBalance);
-  } else {
-    await network.provider.send("hardhat_setBalance", [
-      faucetAddress,
-      "0x313030303030303030303030303030303030303030",
-    ]);
-    await network.provider.request({
-      method: "hardhat_impersonateAccount",
-      params: [faucetAddress],
-    });
-    const faucetSigner = await ethers.provider.getSigner(faucetAddress);
-    await token
-      .connect(faucetSigner)
-      .transfer(targetAddress, await token.balanceOf(faucetAddress));
-
-    await network.provider.request({
-      method: "hardhat_stopImpersonatingAccount",
-      params: [faucetAddress],
-    });
-  }
-};
-
-export const createGauge = async (
-  vaultAddress: string
-): Promise<[Contract, ERC20]> => {
-  const gaugeMockFactory = await ethers.getContractFactory("GaugeMock");
-
-  const gauge = await gaugeMockFactory.deploy(vaultAddress, {
-    gasLimit: 6000000,
-  });
-
-  const stRakisToken = (await ethers.getContractAt(
-    "ERC20",
-    gauge.address
-  )) as ERC20;
-
-  return [gauge, stRakisToken];
-};
-
-export const getArrakisResolver = async (
-  signer: SignerWithAddress
-): Promise<Contract> => {
-  // const resolver = new ethers.Contract(
-  //   addresses.ArrakisV2Resolver,
-  //   ArrakisV2Resolver.abi,
-  //   signer
-  // );
-  const resolver = await ethers.getContractAt(
-    "IArrakisV2Resolver",
-    addresses.ArrakisV2Resolver,
-    signer
-  );
-  return resolver;
 };
